refactor: ルートのPythonスクリプトを整理
- add_prosody_features.py → src/python/piper_train/tools/ に移動
  (`python -m piper_train.tools.add_prosody_features` で実行可能に)
- convert_multi_to_single_speaker.py（ルート）を削除
  (改良版が scripts/ に既に存在するため重複解消)
- テストのimportパスを更新
- CLAUDE.md の便利ツール参照パスを更新
